requirements.txt
# Web Framework
fastapi
uvicorn

# Database
motor
pymongo

# LLM Integration
groq

# Frontend
streamlit

# API Tools
python-multipart
pydantic

# Async Support
asyncio
aiohttp

# Utility
python-dotenv